Added comments to most of the functions and re-indented all of the code

+++ src/Modifiers/VerticalStripes.java
@@ -23,7 +23,7 @@ public class VerticalStripes extends ImageModifier {
         graphics2D.setPaint(new java.awt.Color(0,0,0));
         for(int x = 0 ; x<10; x++){
 
-            graphics2D.fillRect ((int) (x* (input.getImage().getWidth()/10)), (int) 0,(int)(input.getImage().getWidth()/20),(int)(input.getImage().getHeight()));
+            graphics2D.fillRect ((int) (x* (input.getImage().getWidth()/10)),0,(int)(input.getImage().getWidth()/20),(int)(input.getImage().getHeight()));
 
         }
         ImageView returnImageView = input;
